source-code/momento-fiscal-main/etl/requirements.txt
psycopg[binary]>=3.2.0
python-dotenv==1.0.0
tqdm==4.66.1
requests>=2.31.0
googlemaps>=4.10.0
